🙈 Add Gemfile.lock to the gitignore
When reconciling dependencies in a project setting this lockfile is
already ignored. When developers are working on and contributing to
`fake_idp`, the specific dependency lock shouldn't matter - only the
version constraints set in the gemspec.

Ignoring the lockfile and allowing developers to use different
dependencies within the gemspec constraints better simulates the gem's
use in the wild and might enable catching dependency issues sooner